Refactor: 入室時の処理を分離
