+++ discordbotjp/extension.py
@@ -1,5 +1,7 @@
 from discordbotjp.cogs.cog import DiscordBotPortalJP
+from discordbotjp.cogs.join import Join
 
 
 def setup(bot):
     bot.add_cog(DiscordBotPortalJP(bot))
+    bot.add_cog(Join(bot))
